+++ loki-core/src/main/java/io/github/guoshiqiufeng/loki/core/config/BaseCache.java
@@ -19,6 +19,7 @@ import io.github.guoshiqiufeng.loki.core.exception.LokiException;
 import io.github.guoshiqiufeng.loki.core.handler.HandlerHolder;
 import io.github.guoshiqiufeng.loki.core.mapper.BaseMapper;
 import io.github.guoshiqiufeng.loki.core.mapper.BaseMapperImpl;
+import io.github.guoshiqiufeng.loki.support.core.config.LokiProperties;
 import lombok.experimental.UtilityClass;
 
 import java.util.Map;
