+++ loki-core/src/main/java/io/github/guoshiqiufeng/loki/core/handler/impl/KafkaHandler.java
@@ -17,7 +17,8 @@ package io.github.guoshiqiufeng.loki.core.handler.impl;
 
 import io.github.guoshiqiufeng.loki.MessageContent;
 import io.github.guoshiqiufeng.loki.constant.Constant;
-import io.github.guoshiqiufeng.loki.core.config.LokiProperties;
+import io.github.guoshiqiufeng.loki.core.toolkit.StringUtils;
+import io.github.guoshiqiufeng.loki.support.core.config.LokiProperties;
 import io.github.guoshiqiufeng.loki.core.handler.AbstractHandler;
 import io.github.guoshiqiufeng.loki.core.handler.HandlerHolder;
 import io.github.guoshiqiufeng.loki.core.toolkit.KafkaConfigUtils;
@@ -32,7 +33,6 @@ import org.apache.kafka.clients.producer.ProducerRecord;
 import org.apache.kafka.clients.producer.RecordMetadata;
 import org.apache.kafka.common.header.Header;
 import org.apache.kafka.common.header.internals.RecordHeader;
-import org.apache.rocketmq.shaded.commons.lang3.StringUtils;
 
 import java.nio.charset.StandardCharsets;
 import java.util.ArrayList;
@@ -89,7 +89,7 @@ public class KafkaHandler extends AbstractHandler {
         // 发送消息
         try {
             List<Header> headers = new ArrayList<>();
-            if (StringUtils.isNoneBlank(tag)) {
+            if (StringUtils.isNotEmpty(tag)) {
                 headers.add(new RecordHeader(Constant.KAFKA_TAG, tag.getBytes(StandardCharsets.UTF_8)));
             }
             Long timestamp = null;
@@ -134,7 +134,7 @@ public class KafkaHandler extends AbstractHandler {
         // 发送消息
         try {
             List<Header> headers = new ArrayList<>();
-            if (StringUtils.isNoneBlank(tag)) {
+            if (StringUtils.isNotEmpty(tag)) {
                 headers.add(new RecordHeader(Constant.KAFKA_TAG, tag.getBytes(StandardCharsets.UTF_8)));
             }
             Long timestamp = null;
