+++ loki-support/loki-support-core/src/main/java/io/github/guoshiqiufeng/loki/support/core/util/GlobalConfigUtils.java
@@ -13,9 +13,9 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.github.guoshiqiufeng.loki.core.toolkit;
+package io.github.guoshiqiufeng.loki.support.core.util;
 
-import io.github.guoshiqiufeng.loki.core.config.GlobalConfig;
+import io.github.guoshiqiufeng.loki.support.core.config.GlobalConfig;
 import io.github.guoshiqiufeng.loki.enums.MqType;
 import lombok.experimental.UtilityClass;
