+++ loki-support/loki-support-core/src/main/resources/META-INF/spring.factories
@@ -1,0 +1,2 @@
+org.springframework.boot.autoconfigure.EnableAutoConfiguration=\
+  io.github.guoshiqiufeng.loki.support.core.config.LokiPropertiesAutoConfiguration
