+++ loki-support/loki-support-core/src/main/resources/META-INF/spring/org.springframework.boot.autoconfigure.AutoConfiguration.imports
@@ -1,0 +1,1 @@
+io.github.guoshiqiufeng.loki.support.core.config.LokiPropertiesAutoConfiguration
