+++ settings.gradle
@@ -20,6 +20,7 @@ include 'loki-annotation'
 include 'loki-core'
 include 'loki-bom'
 include 'loki-support'
+include 'loki-support:loki-support-core'
 include 'loki-support:loki-support-redis'
 include 'spring-boot-starter'
 include 'spring-boot-starter:loki-spring-boot-autoconfigure'
