+++ spring-boot-starter/loki-spring-boot-autoconfigure/src/main/java/io/github/guoshiqiufeng/loki/autoconfigure/config/LokiAutoConfiguration.java
@@ -17,7 +17,6 @@ package io.github.guoshiqiufeng.loki.autoconfigure.config;
 
 import io.github.guoshiqiufeng.loki.Listener;
 import io.github.guoshiqiufeng.loki.autoconfigure.register.LokiRegistrar;
-import io.github.guoshiqiufeng.loki.core.config.LokiProperties;
 import io.github.guoshiqiufeng.loki.core.exception.LokiException;
 import io.github.guoshiqiufeng.loki.core.handler.Handler;
 import io.github.guoshiqiufeng.loki.core.handler.HandlerHolder;
@@ -26,13 +25,13 @@ import io.github.guoshiqiufeng.loki.core.handler.impl.RedisHandler;
 import io.github.guoshiqiufeng.loki.core.handler.impl.RocketMqHandler;
 import io.github.guoshiqiufeng.loki.core.toolkit.RocketMqConfigUtils;
 import io.github.guoshiqiufeng.loki.enums.MqType;
+import io.github.guoshiqiufeng.loki.support.core.config.LokiProperties;
 import io.github.guoshiqiufeng.loki.support.redis.RedisClient;
 import lombok.extern.slf4j.Slf4j;
 import org.apache.rocketmq.client.apis.ClientException;
 import org.apache.rocketmq.client.apis.producer.Producer;
 import org.springframework.boot.autoconfigure.condition.ConditionalOnMissingBean;
 import org.springframework.boot.autoconfigure.condition.ConditionalOnProperty;
-import org.springframework.boot.context.properties.ConfigurationProperties;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 
@@ -48,34 +47,7 @@ import java.util.List;
  */
 @Slf4j
 @Configuration
-/*@ConditionalOnProperty(prefix = "loki", name = "enabled", matchIfMissing = true)*/
 public class LokiAutoConfiguration {
-
-    /**
-     * 配置文件
-     *
-     * @return 配置文件
-     */
-    @Bean
-    @ConfigurationProperties(prefix = "loki")
-    @ConditionalOnMissingBean(AutoConfigurationProperties.class)
-    public AutoConfigurationProperties lokiProperties() {
-        return new AutoConfigurationProperties();
-    }
-
-    /**
-     * 默认生产者
-     *
-     * @param properties loki配置
-     * @return 默认生产者
-     * @throws ClientException 异常
-     */
-    @Bean
-    @ConditionalOnProperty(prefix = "loki.global-config.mq-config", name = "mq-type", havingValue = "ROCKET_MQ")
-    @ConditionalOnMissingBean(Producer.class)
-    public Producer defaultProducer(LokiProperties properties) throws ClientException {
-        return RocketMqConfigUtils.producerBuilder("defaultProducer", properties);
-    }
 
     /**
      * LokiRegistrar Bean，用于注册处理器和消息监听器
@@ -93,79 +65,4 @@ public class LokiAutoConfiguration {
         return new LokiRegistrar<T>(handlerHolder, lokiProperties, listenerList);
     }
 
-    /**
-     * HandlerHolder Bean，用于持有处理器
-     *
-     * @return HandlerHolder 实例
-     */
-    @Bean
-    public HandlerHolder handlerHolder() {
-        return new HandlerHolder();
-    }
-
-    /**
-     * Handler Bean列表，包含RocketMqHandler
-     *
-     * @param properties    Loki配置
-     * @param handlerHolder 处理器持有者
-     * @return Handler 实例列表
-     */
-    @Bean
-    @ConditionalOnProperty(prefix = "loki.global-config.mq-config", name = "mq-type", havingValue = "ROCKET_MQ")
-    public List<Handler> rocketHandler(LokiProperties properties, HandlerHolder handlerHolder) {
-        ArrayList<Handler> handler = new ArrayList<Handler>(1);
-        if (properties.getGlobalConfig().getMqConfig().getMqType().equals(MqType.ROCKET_MQ)) {
-            RocketMqHandler rocketMqHandler = new RocketMqHandler(properties, handlerHolder);
-            handler.add(rocketMqHandler);
-        } else {
-            throw new LokiException("mq type is not support ");
-        }
-        return handler;
-    }
-
-    /**
-     * Handler Bean列表，包含KafkaHandler
-     *
-     * @param properties    Loki配置
-     * @param handlerHolder 处理器持有者
-     * @return Handler 实例列表
-     */
-    @Bean
-    @ConditionalOnProperty(prefix = "loki.global-config.mq-config", name = "mq-type", havingValue = "KAFKA")
-    public List<Handler> kafkaHandler(LokiProperties properties, HandlerHolder handlerHolder) {
-        ArrayList<Handler> handler = new ArrayList<Handler>(1);
-        if (properties.getGlobalConfig().getMqConfig().getMqType().equals(MqType.KAFKA)) {
-            KafkaHandler kafkaHandler = new KafkaHandler(properties, handlerHolder);
-            handler.add(kafkaHandler);
-        } else {
-            throw new LokiException("mq type is not support ");
-        }
-        return handler;
-    }
-
-    /**
-     * Handler Bean列表，包含RedisHandler
-     *
-     * @param properties    Loki配置
-     * @param handlerHolder 处理器持有者
-     * @return Handler 实例列表
-     */
-    @Bean
-    @ConditionalOnProperty(prefix = "loki.global-config.mq-config", name = "mq-type", havingValue = "REDIS")
-    public List<Handler> redisHandler(LokiProperties properties, HandlerHolder handlerHolder, RedisClient redisClient) {
-        ArrayList<Handler> handler = new ArrayList<Handler>(1);
-        if (properties.getGlobalConfig().getMqConfig().getMqType().equals(MqType.REDIS)) {
-            RedisHandler redisHandler = new RedisHandler(properties, handlerHolder, redisClient);
-            handler.add(redisHandler);
-        } else {
-            throw new LokiException("mq type is not support ");
-        }
-        return handler;
-    }
-
-    /**
-     * 构造函数
-     */
-    public LokiAutoConfiguration() {
-    }
 }
