+++ spring-boot-starter/loki-spring-boot-autoconfigure/src/main/resources/META-INF/spring.factories
@@ -1,2 +1,5 @@
 org.springframework.boot.autoconfigure.EnableAutoConfiguration=\
+  io.github.guoshiqiufeng.loki.autoconfigure.config.RocketMqAutoConfiguration,\
+  io.github.guoshiqiufeng.loki.autoconfigure.config.KafkaAutoConfiguration,\
+  io.github.guoshiqiufeng.loki.autoconfigure.config.RedisAutoConfiguration,\
   io.github.guoshiqiufeng.loki.autoconfigure.config.LokiAutoConfiguration
