+++ spring-boot-starter/loki-spring-boot-autoconfigure/src/main/resources/META-INF/spring/org.springframework.boot.autoconfigure.AutoConfiguration.imports
@@ -1,1 +1,4 @@
+io.github.guoshiqiufeng.loki.autoconfigure.config.RocketMqAutoConfiguration
+io.github.guoshiqiufeng.loki.autoconfigure.config.KafkaAutoConfiguration
+io.github.guoshiqiufeng.loki.autoconfigure.config.RedisAutoConfiguration
 io.github.guoshiqiufeng.loki.autoconfigure.config.LokiAutoConfiguration
